+++ run/kubejs/startup_scripts/item/misc.js
@@ -3,4 +3,8 @@ StartupEvents.registry('item', e => {
         .bladeColor(0x545F977);
     e.create('glassworking', 'tfc:glassworking')
         .operation('kubejs_quench');
+    e.create('jar', 'tfc:jar')
+        .placedModel('kubejs:block/test_jar');
+    e.create('topped_jar', 'tfc:jar')
+        .placedModel('kubejs:block/topped_jar');
 })